Refactor worker code into its own module.
